renamed folders, implemented ball disappearing when you get relic
