+++ src/main/resources/anniv7Resources/images/characters/Dandadan/Ball.png
