+++ src/main/resources/anniv7Resources/images/characters/Dandadan/BallPortrait.png
