+++ src/main/resources/anniv7Resources/images/characters/Dandadan/acrobatic_silky.png
