+++ src/main/resources/anniv7Resources/images/characters/Dandadan/ghost.png
