+++ src/main/resources/anniv7Resources/images/characters/Dandadan/turbo_granny.png
