+++ src/main/resources/anniv7Resources/localization/eng/DandadanPatron/Cutscenestrings.json
@@ -8,7 +8,7 @@
       "\"So which will it be?\"",
       "Suddenly you feel whole again, as if you had regained something you ~lost~ .",
       "\"Enjoy a ~taste~ of my power.\"",
-      "You inspect the card... it seems to be autographed bootleg merch.",
+      "You inspect the card... it seems to be an autographed bootleg.",
       "\"I was saving that one for someone special but I guess you can have it as well... ~hehehe~ .\"",
       "You pocket a ~slimy~ potion vial that looks suspiciously like the ball you got it from...",
       "\"Put that to good use, took me a while to save up that much.\"",
